requirements-integration.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
pydantic>=2.0
python-dotenv
cryptography
httpx[socks]
watchdog
celery
redis
psycopg2-binary
asyncpg
orjson
prometheus-client
pytest
pytest-asyncio

